# server/docs/swagger.yaml
definitions:
  models.Todo:
    properties:
      body:
        type: string
      category:
        type: string
      deadline:
        type: string
      done:
        type: boolean
      id:
        type: integer
      title:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/todos:
    get:
      description: Get a list of all todos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Todo'
            type: array
      summary: Get all todos
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Create a new todo item and store it in the database
      parameters:
      - description: Todo
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/models.Todo'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Todo'
      summary: Create a new todo
      tags:
      - todos
  /api/todos/{id}:
    delete:
      description: Delete a specific todo item by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete a todo item
      tags:
      - todos
    patch:
      consumes:
      - application/json
      description: Update the details of an existing todo item
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Todo
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/models.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
      summary: Update a todo item
      tags:
      - todos
  /api/todos/{id}/done:
    patch:
      description: Toggle the "done" status of a specific todo item
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Toggle todo completion status
      tags:
      - todos
swagger: "2.0"